+++ greenfield/eldenring/tests/test_gf_option_template_yaml.py
@@ -158,7 +158,7 @@ def test_the_generated_template_keeps_every_default_weighted():
 def test_generated_template_omits_advanced_and_compatibility_only_controls():
     block = Utils.parse_yaml(_generate_template_for_this_game())[GAME]
     omitted = {"vanilla_placement", "flask_upgrades_on_progression_surface",
-               "global_scadutree_blessing", "merchant_bell_logic"}
+               "global_scadutree_blessing", "merchant_bell_logic", "leyndell_runes_required"}
     classes = dict(_option_classes())
     assert omitted <= classes.keys()
     assert not omitted & block.keys()